leak-finder/requirements.txt
gitpython>=3.1.0
pydantic>=2.0.0
openai>=1.0.0
anthropic>=0.18.0
google-genai>=1.0.0
orchagent-sdk>=0.1.0
